Enable configuration (#5)
This uses Pydantic settings to load settings from environment variables. For example, in `make run`, we use a `.env` file to populate the variables and then can run the program. This approach is very dynamic and will work with docker compose and kubernetes deployments if needed.

We also use the prebound method pattern. This pattern decouples the configuration from the rest of the code, as it only loads environment variables when using setup() or accessing config attributes.

Further reading:
- [The prebound method pattern](https://python-patterns.guide/python/prebound-methods/)

---

Squashed commits:

* configure using prebound pattern

* add logs and dynamic config access

* validate delayed configuration

* improve after review